Reformatted (cleaned) entire solution structure to match the structure of master branch
