--- SecretSanta/SecretSanta.Web.Tests/UnitTest1.cs
@@ -1,13 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace SecretSanta.Web.Tests
-{
-    [TestClass]
-    public class UnitTest1
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-        }
-    }
-}
