+++ SecretSanta/SecretSanta.sln
@@ -13,23 +13,21 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Api", "src\Secr
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Web", "src\SecretSanta.Web\SecretSanta.Web.csproj", "{E5663FEA-B7B2-4C84-B6F7-44744A427C87}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Business.Tests", "src\SecretSanta.BusinessTests\SecretSanta.Business.Tests.csproj", "{021FB7D2-773B-4C86-9B35-60FE2A683D37}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B3545995-45E9-4843-A297-736B57ED2884}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{1A817B7D-1E5A-4ED7-8DBB-CF448CC2C976}"
 	ProjectSection(SolutionItems) = preProject
 		Directory.Build.props = Directory.Build.props
+		GlobalSuppression.cs = GlobalSuppression.cs
 	EndProjectSection
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{2FB717F7-CA90-4442-84FC-570AC94D8174}"
-	ProjectSection(SolutionItems) = preProject
-		src\SecretSanta.BusinessTests\TestSuppressions.cs = src\SecretSanta.BusinessTests\TestSuppressions.cs
-	EndProjectSection
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{96D1C573-350A-4E65-BBD1-415BA8482E43}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Api.Tests", "SecretSanta.Api.Tests\SecretSanta.Api.Tests.csproj", "{58379FAA-6E45-4B55-BE66-1A4559CB94BC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretSanta.Api.Tests", "test\SecretSanta.Api.Tests\SecretSanta.Api.Tests.csproj", "{44F7A03C-672B-452D-9943-76729BFDA3EF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Data.Tests", "SecretSanta.Data.Tests\SecretSanta.Data.Tests.csproj", "{5A19CDD6-ADEC-4958-AC94-A53A56A4009E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretSanta.Business.Tests", "test\SecretSanta.Business.Tests\SecretSanta.Business.Tests.csproj", "{5A25FC10-6F72-4311-9AE1-AF13EBB19E02}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SecretSanta.Web.Tests", "SecretSanta.Web.Tests\SecretSanta.Web.Tests.csproj", "{2F375CF5-53CC-45DD-9583-8D14F9A8314A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretSanta.Data.Tests", "test\SecretSanta.Data.Tests\SecretSanta.Data.Tests.csproj", "{47D92C37-5A60-4F42-9524-21698B470C3D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretSanta.Web.Tests", "test\SecretSanta.Web.Tests\SecretSanta.Web.Tests.csproj", "{297E6558-9B44-4359-BD44-DFFEF0C2C6BF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,22 +51,22 @@ Global
 		{E5663FEA-B7B2-4C84-B6F7-44744A427C87}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E5663FEA-B7B2-4C84-B6F7-44744A427C87}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E5663FEA-B7B2-4C84-B6F7-44744A427C87}.Release|Any CPU.Build.0 = Release|Any CPU
-		{021FB7D2-773B-4C86-9B35-60FE2A683D37}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{021FB7D2-773B-4C86-9B35-60FE2A683D37}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{021FB7D2-773B-4C86-9B35-60FE2A683D37}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{021FB7D2-773B-4C86-9B35-60FE2A683D37}.Release|Any CPU.Build.0 = Release|Any CPU
-		{58379FAA-6E45-4B55-BE66-1A4559CB94BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{58379FAA-6E45-4B55-BE66-1A4559CB94BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{58379FAA-6E45-4B55-BE66-1A4559CB94BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{58379FAA-6E45-4B55-BE66-1A4559CB94BC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5A19CDD6-ADEC-4958-AC94-A53A56A4009E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5A19CDD6-ADEC-4958-AC94-A53A56A4009E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5A19CDD6-ADEC-4958-AC94-A53A56A4009E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5A19CDD6-ADEC-4958-AC94-A53A56A4009E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2F375CF5-53CC-45DD-9583-8D14F9A8314A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2F375CF5-53CC-45DD-9583-8D14F9A8314A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2F375CF5-53CC-45DD-9583-8D14F9A8314A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2F375CF5-53CC-45DD-9583-8D14F9A8314A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{44F7A03C-672B-452D-9943-76729BFDA3EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{44F7A03C-672B-452D-9943-76729BFDA3EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{44F7A03C-672B-452D-9943-76729BFDA3EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{44F7A03C-672B-452D-9943-76729BFDA3EF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5A25FC10-6F72-4311-9AE1-AF13EBB19E02}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5A25FC10-6F72-4311-9AE1-AF13EBB19E02}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5A25FC10-6F72-4311-9AE1-AF13EBB19E02}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5A25FC10-6F72-4311-9AE1-AF13EBB19E02}.Release|Any CPU.Build.0 = Release|Any CPU
+		{47D92C37-5A60-4F42-9524-21698B470C3D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{47D92C37-5A60-4F42-9524-21698B470C3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{47D92C37-5A60-4F42-9524-21698B470C3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{47D92C37-5A60-4F42-9524-21698B470C3D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{297E6558-9B44-4359-BD44-DFFEF0C2C6BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{297E6558-9B44-4359-BD44-DFFEF0C2C6BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{297E6558-9B44-4359-BD44-DFFEF0C2C6BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{297E6558-9B44-4359-BD44-DFFEF0C2C6BF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -78,10 +76,10 @@ Global
 		{1CFA6F17-4CD4-4BB7-BF5C-C1704ADCA86B} = {71196D9A-D1FA-469B-86BD-08620BD5BC85}
 		{C33A221C-CDAB-4245-A9A4-B32A5AB9A690} = {71196D9A-D1FA-469B-86BD-08620BD5BC85}
 		{E5663FEA-B7B2-4C84-B6F7-44744A427C87} = {71196D9A-D1FA-469B-86BD-08620BD5BC85}
-		{021FB7D2-773B-4C86-9B35-60FE2A683D37} = {2FB717F7-CA90-4442-84FC-570AC94D8174}
-		{58379FAA-6E45-4B55-BE66-1A4559CB94BC} = {2FB717F7-CA90-4442-84FC-570AC94D8174}
-		{5A19CDD6-ADEC-4958-AC94-A53A56A4009E} = {2FB717F7-CA90-4442-84FC-570AC94D8174}
-		{2F375CF5-53CC-45DD-9583-8D14F9A8314A} = {2FB717F7-CA90-4442-84FC-570AC94D8174}
+		{44F7A03C-672B-452D-9943-76729BFDA3EF} = {96D1C573-350A-4E65-BBD1-415BA8482E43}
+		{5A25FC10-6F72-4311-9AE1-AF13EBB19E02} = {96D1C573-350A-4E65-BBD1-415BA8482E43}
+		{47D92C37-5A60-4F42-9524-21698B470C3D} = {96D1C573-350A-4E65-BBD1-415BA8482E43}
+		{297E6558-9B44-4359-BD44-DFFEF0C2C6BF} = {96D1C573-350A-4E65-BBD1-415BA8482E43}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E2234040-21B9-4AF6-9A6A-3ABE8D5F86E0}
