+++ SecretSanta/src/SecretSanta.Api/Program.cs
@@ -1,14 +1,19 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
 namespace SecretSanta.Api
 {
-    public static class Program
+    public class Program
     {
         public static void Main(string[] args)
         {
             CreateHostBuilder(args).Build().Run();
-    
         }
 
         public static IHostBuilder CreateHostBuilder(string[] args) =>
@@ -17,6 +22,5 @@ namespace SecretSanta.Api
                 {
                     webBuilder.UseStartup<Startup>();
                 });
-       
     }
 }
