+++ SecretSanta/src/SecretSanta.Api/SecretSanta.Api.csproj
@@ -6,15 +6,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="IntelliTect.Analyzers" Version="0.1.5" />
+    <Compile Include="..\..\GlobalSuppression.cs" Link="GlobalSuppression.cs" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\SecretSanta.Business\SecretSanta.Business.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Include="..\GlobalSuppressions.cs" Link="GlobalSuppressions.cs" />
   </ItemGroup>
 
 </Project>
