+++ SecretSanta/src/SecretSanta.Business/SecretSanta.Business.csproj
@@ -1,14 +1,15 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
-
+  <ItemGroup>
+    <Compile Include="..\..\GlobalSuppression.cs" Link="GlobalSuppression.cs" />
+  </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\SecretSanta.Data\SecretSanta.Data.csproj" />
-    <Compile Include="..\..\GlobalSuppression.cs" Link="GlobalSuppression.cs" />
   </ItemGroup>
 
 </Project>
