+++ SecretSanta/src/SecretSanta.Data/SecretSanta.Data.csproj
@@ -8,6 +8,10 @@
     <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.1" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.1.1" />
+    
+  </ItemGroup>
+
+    <ItemGroup>
     <Compile Include="..\..\GlobalSuppression.cs" Link="GlobalSuppression.cs" />
   </ItemGroup>
 </Project>
