+++ SecretSanta/src/SecretSanta.Web/Program.cs
@@ -1,9 +1,15 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
 namespace SecretSanta.Web
 {
-    public static class Program
+    public class Program
     {
         public static void Main(string[] args)
         {
