+++ SecretSanta/test/SecretSanta.Business.Tests/SecretSanta.Business.Tests.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <Compile Include="..\..\GlobalSuppression.cs" Link="GlobalSuppression.cs" />
+    <Compile Include="..\..\test\TestSuppression.cs" Link="TestSuppression.cs" />
   </ItemGroup>
 
   <ItemGroup>
